LayoutTests/java/lc3/ConvertJSObject/ToShort-001-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  JavaScript Object to short

FAIL TEST_CLASS.ambiguous( new String() ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Boolean() ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Number() ) +'' should be SHORT. Was BYTE.
FAIL TEST_CLASS.ambiguous( new Date(999) ) +'' should be SHORT. Was BYTE.
PASS successfullyParsed is true

TEST COMPLETE
 
